@@ -4,7 +4,8 @@
     "type": "mautic-plugin",
     "version": "1.0.0",
     "require": {
-        "php": ">=7.4",
+        "php": ">=8.1",
+        "mautic/core-lib": "^6.0",
         "stripe/stripe-php": "^10.0"
     },
     "autoload": {
